election_patrol_backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo[srv]==4.7.1
motor==3.4.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic[email]==2.7.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
httpx==0.27.0
firebase-admin==6.5.0
